@@ -3,6 +3,9 @@ package br.com.casadocodigo.controller;
 import java.util.concurrent.Callable;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.MailSender;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -13,6 +16,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import br.com.casadocodigo.model.CarrinhoCompras;
 import br.com.casadocodigo.model.DadosPagamento;
+import br.com.casadocodigo.model.Usuario;
 
 @Controller
 @RequestMapping("/pagamento")
@@ -24,32 +28,39 @@ public class PagamentoController {
 	@Autowired
 	RestTemplate restTemplate;
 
+	@Autowired
+	private MailSender sender;
+
 	@RequestMapping(value = "/finalizar", method = RequestMethod.POST)
-	public Callable<ModelAndView> finalizar(RedirectAttributes model) {
-		return new Callable<ModelAndView>() {
-			private RedirectAttributes model;
+	public Callable<ModelAndView> finalizar(@AuthenticationPrincipal Usuario usuario, RedirectAttributes model) {
+		return () -> {
+			String uri = "http://book-payment.herokuapp.com/payment";
 
-			public Callable<ModelAndView> setModel(RedirectAttributes model) {
-				this.model = model;
-				return this;
+			try {
+				String response = restTemplate.postForObject(uri, new DadosPagamento(carrinho.getTotal()),
+						String.class);
+				model.addFlashAttribute("message", response);
+
+				enviaEmailPedido(usuario);
+
+			} catch (HttpClientErrorException e) {
+				e.printStackTrace();
+				model.addFlashAttribute("message", "Valor maior que o permitido");
 			}
 
-			@Override
-			public ModelAndView call() throws Exception {
-				ModelAndView modelAndView = new ModelAndView("redirect:/produtos");
-				String response;
-				try {
-					String uri = "http://book-payment.herokuapp.com/payment";
-					response = restTemplate.postForObject(uri, new DadosPagamento(carrinho.getTotal()), String.class);
-					model.addFlashAttribute("message", response);
-				} catch (HttpClientErrorException e) {
-					e.printStackTrace();
-					model.addFlashAttribute("message", "Valor maior que o permitido");
-				}
-				return modelAndView;
-			}
-
-		}.setModel(model);
+			return new ModelAndView("redirect:/produtos");
+		};
 
 	}
+
+	private void enviaEmailPedido(Usuario usuario) {
+		SimpleMailMessage mailMessage = new SimpleMailMessage();
+		mailMessage.setFrom("webapp.java.test@gmail.com");
+		mailMessage.setTo("webapp.java.test@gmail.com"); // usuario.getEmail();
+		mailMessage.setSubject("Compra finalizada com Sucesso!");
+		mailMessage.setText("Compra aprovada no valor de R$" + carrinho.getTotal());
+
+		sender.send(mailMessage);
+	}
+
 }
