@@ -3,8 +3,10 @@ package br.com.casadocodigo.configuration;
 import java.util.Properties;
 
 import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
 
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Profile;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.JpaVendorAdapter;
@@ -15,31 +17,35 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @EnableTransactionManagement
 public class JPAConfiguration {
 	@Bean
-	public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+	public LocalContainerEntityManagerFactoryBean entityManagerFactory(DataSource dataSource) {
 		LocalContainerEntityManagerFactoryBean factoryBean = new LocalContainerEntityManagerFactoryBean();
+		factoryBean.setPackagesToScan("br.com.casadocodigo.model");
+		factoryBean.setDataSource(dataSource);
 
 		JpaVendorAdapter jpaVendorAdapter = new HibernateJpaVendorAdapter();
-
 		factoryBean.setJpaVendorAdapter(jpaVendorAdapter);
+		factoryBean.setJpaProperties(aditionalProperties());
 
+		return factoryBean;
+	}
+
+	private Properties aditionalProperties() {
+		Properties props = new Properties();
+		props.setProperty("hibernate.dialect", "org.hibernate.dialect.MySQL5Dialect");
+		props.setProperty("hibernate.show_sql", "true");
+		props.setProperty("hibernate.hbm2ddl.auto", "update");
+		return props;
+	}
+
+	@Bean
+	@Profile("dev")
+	public DataSource dataSource() {
 		DriverManagerDataSource dataSource = new DriverManagerDataSource();
 		dataSource.setUsername("root");
 		dataSource.setPassword("root");
 		dataSource.setUrl("jdbc:mysql://localhost:3306/casadocodigo");
 		dataSource.setDriverClassName("com.mysql.jdbc.Driver");
-
-		factoryBean.setDataSource(dataSource);
-
-		Properties props = new Properties();
-		props.setProperty("hibernate.dialect", "org.hibernate.dialect.MySQL5Dialect");
-		props.setProperty("hibernate.show_sql", "true");
-		props.setProperty("hibernate.hbm2ddl.auto", "update");
-
-		factoryBean.setJpaProperties(props);
-
-		factoryBean.setPackagesToScan("br.com.casadocodigo.model");
-
-		return factoryBean;
+		return dataSource;
 	}
 
 	@Bean
